requirements.txt
alabaster==0.7.12
anyio==3.6.1
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
brotlipy==0.7.0
cachetools==4.2.2
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.0.5
coverage==6.3.2
cryptography==37.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docutils==0.18.1
dropbox==11.35.0
dropboxdrivefs==1.3.1
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.37.4
fsspec==2022.8.2
funcy==1.17
gitdb==4.0.9
GitPython==3.1.29
httplib2==0.20.4
idna==3.3
imagesize==1.4.1
Jinja2==3.0.3
json5==0.9.10
jsonschema==4.16.0
kiwisolver==1.4.4
loguru==0.6.0
MarkupSafe==2.1.1
mistune==2.0.4
nest-asyncio==1.5.6
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyaml==21.10.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0b4
PyDrive2==1.14.0
pyflakes==2.4.0
Pygments==2.11.2
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==13.14
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-cache==0.9.6
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
semver==2.13.0
Send2Trash==1.8.0
setuptools==63.4.1
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==5.4.0
txt2tags==3.7
typer==0.6.1
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.11
validators==0.20.0
watchdog==2.1.9
webencodings==0.5.1
websocket-client==1.4.1
wheel==0.37.1
widgetsnbextension==4.0.3
zipp==3.8.0
